pulled permissions form into its own control and used that in both location and device permissions
