+++ src/lib/services/LocationService.ts
@@ -244,11 +244,13 @@ export class LocationService implements ILocationService {
 	 * @param locationId The location ID
 	 * @param userId The user ID to remove
 	 * @param locationOwnerId The location owner entry ID
+	 * @param me The ID of the user performing the action (to check if they are removing themselves)
 	 */
 	async removeUserFromLocation(
 		locationId: number,
 		userId: string,
-		locationOwnerId: number
+		locationOwnerId: number,
+		me: string
 	): Promise<{ success: boolean; error?: string }> {
 		try {
 			// Check if user is attempting to remove themselves as admin
@@ -266,7 +268,7 @@ export class LocationService implements ILocationService {
 				}
 			}
 
-			await this.locationRepository.removeUserFromLocationByUserId(userId);
+			await this.locationRepository.removeUserFromLocationByUserId(userId, locationId);
 			// The user no longer has access to this location, thus they will no longer have access to any devices in this location.
 			// Delete the permission to all devices for this user too
